// apps/organization/src/app/departments/enums/DepartmentEnums.ts
export enum departmentEnum {
  LOADING = 'Laoding',
  ERROR = 'Error',
  DEPARTMENTS = 'Departments',
  NEW_DEPARTMENT = 'New Department',
  ASC = 'ASC',
  DESC = 'DESC',
  Name = 'name',
  CREATED_AT = 'createdAt',
  NO_RECORDS = 'No Records...',
  LABEL_PROJECT = 'Project',
  LABEL_CLIENT = 'Client',
  PLACE_HOLDER_CLIENT = '--Client--',
  PLACE_HOLDER_PROJECT = '--Project--',
  PLACE_HOLDER_SEARCH = 'Search',
  AVATAR = 'Avatar',
  ARCHIVE="Archive",
  DELETE='Delete',
  EDIT='Edit',
  DELETE_MESSAGE="Do you really want to delete this department?",
  ARCHIVE_MESSAGE="Do you really want to archive this department?",
  NEXT = 'Next',
  PREVIOUS = 'Previous',
  ENTRIES="Entries",
    
}
export enum addDepartmentEnum {
  ADD_NEW_DEPARTMENT = 'Add New Department',
  EDIT_DEPARTMENT = 'Edit Department',
  BACK = 'Back',
  SAVE = 'Save',
  CANCEL = 'Cancel',
  EDIT = 'Update',
  DEPARTMENT_NAME = 'Department Name',
  SELECTED_DEPARTMENT = 'Select departments',
  USERS = 'Users',
  SELECT_CLIENT = 'Select Client',
  CLIENT = 'Client',
  INCLUDING_MEMBERS = 'Including Members',
  DESCRIPTION = 'Description',
  ID_DESCRIPTION = 'description',
  NAME = 'name',
}
export enum archieveEnum {
  NO_RECORDS = 'No records...',
  DEPARTMENTS = 'Departments',
  NEW_DEPARTMENT = 'New Department',
  UN_ARCHIVE_MESSAGE = 'Do you really want to un-archive this department?',
  DELETE_MESSAGE = 'Do you really want to delete this department?',
  NEXT = 'Next',
  PREVIOUS = 'Previous',
  PAGE_SIZE = 'Page Size',
  AVATAR = 'Avatar',
  ICON = 'icon',
  DELETE = 'Delete',
  UNARCHIVE = 'Un-archive',
  ARCHIVE = 'Archive',
  ENTRIES = 'Entries',
  CLIENT="Client",
  PROJECT="Project",
  PLACE_HOLDER_CLIENT = '--Client--',
  PLACE_HOLDER_PROJECT = '--Project--',
}